pyproject.toml
[project]
name = "sustainable-mode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.37.13",
    "browser-use>=0.1.40",
    "langchain>=0.3.20",
    "langchain-aws>=0.2.15",
    "langchain-community>=0.3.19",
    "markdown>=3.7",
    "python-dotenv>=1.0.1",
]
